Update index.html
Updated index.html by adding a content container
to contain all the other elements beside icon.
This is to arrange the warning icon and the
container and its contents into rows.

Inside the content container. The header and close
button are within in alert-close container. This
is to horizontally align the header and close
button while maintaining spaces between each
other.

The rest are the same.