Update style.css
Updated style.css by adding my solutions. Made
body as flex container to arrange the three
sections into column.

Grouped header and footer together to apply as
flex container and distribute flex items within it
as space-between horizontally. Added some paddings
as well.

Removed the bullet in ul and made it as flex
container as well to arrange the items as rows.

Removed text decorations and color in anchor tags.

The others are adjusted slightly.